+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.1
+- Small playtesting shows that the default beacon cooldowns are too low, so they have been increased by a bit.
+  - Healing + Shocking Beacon Cooldowns: 40s -> 60s
+  - Equipment Beacon Cooldown: 60s -> 90s
+  - hacking Beacon Cooldown: 180s -> 240s
+  - Of course, you can change these in the configs still.
+
+- As of the second DLC's release, the beacons seem to be resetting sooner than they should be.
+It should be fixed as of this version.
+
+- Backup mags used to (unintentionally) increase the max number of charges of the right beacon. This has now been fixed this update
+
 ## 1.0.0
 
 - Initial release